默认封包解包实现

+++ niface/Inotify.go
@@ -2,7 +2,7 @@
  * @Author: liusuxian 382185882@qq.com
  * @Date: 2023-02-19 01:40:25
  * @LastEditors: liusuxian 382185882@qq.com
- * @LastEditTime: 2023-03-08 14:52:42
+ * @LastEditTime: 2023-03-15 18:23:19
  * @FilePath: /playlet-server/Users/liusuxian/Desktop/project-code/golang-project/nova/niface/Inotify.go
  * @Description:
  *
@@ -17,8 +17,8 @@ type INotify interface {
 	AddNotify(id uint64, conn IConnection)                       // 添加通知连接
 	GetNotify(id uint64) (IConnection, error)                    // 获取通知连接
 	DelNotify(id uint64)                                         // 删除通知连接
-	NotifyToConn(id uint64, msgID uint32, data []byte) error     // 通知某个连接
-	NotifyAll(msgID uint32, data []byte) error                   // 通知所有连接
-	NotifyBuffToConn(id uint64, msgID uint32, data []byte) error // 通过缓冲队列通知某个连接
-	NotifyBuffAll(msgID uint32, data []byte) error               // 通过缓冲队列通知所有连接
+	NotifyToConn(id uint64, msgID uint16, data []byte) error     // 通知某个连接
+	NotifyAll(msgID uint16, data []byte) error                   // 通知所有连接
+	NotifyBuffToConn(id uint64, msgID uint16, data []byte) error // 通过缓冲队列通知某个连接
+	NotifyBuffAll(msgID uint16, data []byte) error               // 通过缓冲队列通知所有连接
 }
