+++ niface/iserver.go
@@ -2,7 +2,7 @@
  * @Author: liusuxian 382185882@qq.com
  * @Date: 2023-02-18 23:25:31
  * @LastEditors: liusuxian 382185882@qq.com
- * @LastEditTime: 2023-03-15 00:18:47
+ * @LastEditTime: 2023-03-15 17:39:47
  * @FilePath: /playlet-server/Users/liusuxian/Desktop/project-code/golang-project/nova/niface/iserver.go
  * @Description:
  *
@@ -14,7 +14,7 @@ package niface
 type IServer interface {
 	Start()                                 // 启动服务器
 	Stop()                                  // 停止服务器
-	AddRouter(msgID uint32, router IRouter) // 路由功能：给当前服务注册一个路由业务方法，供客户端连接处理使用
+	AddRouter(msgID uint16, router IRouter) // 路由功能：给当前服务注册一个路由业务方法，供客户端连接处理使用
 	GetConnManager() IConnManager           // 获取连接管理
 	SetOnConnStart(func(IConnection))       // 设置当前 Server 的连接创建时的 Hook 函数
 	SetOnConnStop(func(IConnection))        // 设置当前 Server 的连接断开时的 Hook 函数
